Ignore underscores in `declare()`.
Seems like version.pm is a bit inconsistent when a version string is passed
that has underscores in it. So let it do its thing in `parse()`, but strip
them out in `declare()`. And even if it retains them in `parse()`, strip them
out and clean things up in `normal()`.