Fix declare() with dash and extra
SemVer has a bug introduced in 8781c0d987f2abde08293a7ea127b68167bc9cb6:

  $ perl -MSemVer -le'print SemVer->declare("3.1.0")->normal'
  3.1.0

  $ perl -MSemVer -le'print SemVer->declare("3.1.0-something")->normal'
  Invalid version format (version required) at .../SemVer.pm line 90.

What happens in the second case ("3.1.0-something") is that this regular
expression matches:

  (my $v = $ival) =~ s/^v?$STRICT_DOTTED_INTEGER_VERSION(?:($DASH_SEPARATOR)($OPTIONAL_EXTRA_PART))[[:space:]]*$//;

So the entire version string is replaced by nothing, and that is the
value that $v gets. All other cases (such as "3.1", or "3.1.0") don't
match, and so the replacement never happens.

This empty $v is then passed on to version->declare, which complains
that the version is required.

It seems the full regexp was added hoping to be a validation step for
strict semantic versions in declare() and parse(). Unfortunately, this
wasn't working (as explained above), and in fact, that is not the
desired behaviour for these functions. From the docs for declare():
"This is the most flexible way to declare versions. Consider using it to
normalize version strings." With that in mind, the input has to allow
"not normal" version strings.

This patch restores the old behaviour, fixing usage of declare() and
parse(). I'm not super comfortable with how parse() is supposed to work,
so I didn't add any tests for it. But I did add tests for declare().